+++ public/js/cat_source/es6/components/segments/Segment.js
@@ -20,6 +20,7 @@ import ConfirmMessageModal from '../modals/ConfirmMessageModal'
 import SegmentBody from './SegmentBody'
 import TranslationIssuesSideButton from '../review/TranslationIssuesSideButton'
 import MBC from '../../utils/mbc.main'
+import ModalsActions from '../../actions/ModalsActions'
 
 class Segment extends React.Component {
   constructor(props) {
@@ -72,9 +73,9 @@ class Segment extends React.Component {
     if (!this.$section.length) return
     if (!this.checkIfCanOpenSegment()) {
       if (UI.projectStats && UI.projectStats.TRANSLATED_PERC_FORMATTED === 0) {
-        alertNoTranslatedSegments()
+        this.alertNoTranslatedSegments()
       } else {
-        alertNotTranslatedYet(this.props.segment.sid)
+        this.alertNotTranslatedYet(this.props.segment.sid)
       }
     } else {
       if (this.props.segment.translation.length !== 0) {
@@ -119,6 +120,30 @@ class Segment extends React.Component {
         document.location.pathname + '#' + this.props.segment.sid,
       )
     }
+  }
+
+  alertNotTranslatedYet = (sid) => {
+    setTimeout(() =>
+      ModalsActions.showModalComponent(ConfirmMessageModal, {
+        cancelText: 'Close',
+        successCallback: () => UI.openNextTranslated(sid),
+        successText: 'Open next translated segment',
+        text: UI.alertNotTranslatedMessage,
+      }),
+    )
+  }
+
+  alertNoTranslatedSegments = () => {
+    var props = {
+      text: 'There are no translated segments to revise in this job.',
+      successText: 'Ok',
+      successCallback: function () {
+        ModalsActions.onCloseModal()
+      },
+    }
+    setTimeout(() =>
+      ModalsActions.showModalComponent(ConfirmMessageModal, props, 'Warning'),
+    )
   }
 
   openSegmentFromAction(sid) {
